Fix Fuentes int input truncation by forcing float dtype

+++ pvlib/temperature.py
@@ -884,7 +884,7 @@ def fuentes(poa_global, temp_air, wind_speed, noct_installed, module_height=5,
     windmod_array = wind_speed * (module_height/wind_height)**0.2 + 1e-4
 
     tmod0 = 293.15
-    tmod_array = np.zeros_like(poa_global)
+    tmod_array = np.zeros_like(poa_global, dtype = float)
 
     iterator = zip(tamb_array, sun_array, windmod_array, tsky_array,
                    timedelta_hours)
